I chaged it to double

+++ CS597-PROJECT/src/TestFile.java
@@ -1,5 +1,5 @@
 
 public class TestFile {
 	int x=10;
-	float y=10;
+	double y=10.0 ;
 }
